chore(scripts): move scripts to their own folders
