+++ package.json
@@ -22,13 +22,13 @@
     "start": "yarn run dev",
     "dev": "yarn run storybook",
     "build": "yarn run build:cjs && yarn run build:es && yarn run build:umd",
-    "build:umd": "BABEL_ENV=umd rollup --config scripts/rollup.config.js",
+    "build:umd": "BABEL_ENV=umd rollup --config scripts/rollup/config.js",
     "build:cjs": "BABEL_ENV=cjs babel src --extensions '.js,.ts,.tsx' --out-dir cjs/ --ignore 'src/index.es.js','**/__tests__','**/__mocks__' --quiet",
     "build:es": "BABEL_ENV=es babel src --extensions '.js,.ts,.tsx' --out-dir es/ --ignore 'src/index.es.js','**/__tests__','**/__mocks__' --quiet && BABEL_ENV=es babel src/index.es.js --out-file es/index.js --quiet",
     "doctoc": "doctoc --no-title --maxlevel 3 README.md CONTRIBUTING.md",
     "storybook": "start-storybook -p 6006 --ci --static-dir .storybook/static",
     "storybook:build": "build-storybook --output-dir website/stories --static-dir .storybook/static",
-    "website:build": "yarn run storybook:build && ./scripts/zip-examples.sh",
+    "website:build": "yarn run storybook:build && ./scripts/website/zip-examples.sh",
     "lint": "eslint . && tslint --project .",
     "lint:fix": "eslint . --fix && tslint --project . --fix",
     "type-check": "tsc",
@@ -37,11 +37,11 @@
     "test:watch": "jest --config scripts/jest/config.js --watch --bail",
     "test:size": "bundlesize",
     "test:functional": "wdio functional-tests/wdio.conf.js",
-    "test:functional:dev": "babel-node scripts/dev-functional-tests",
-    "test:functional:dev:debug-server": "PORT=9090 babel-node scripts/dev-functional-tests-debug-server",
+    "test:functional:dev": "babel-node scripts/functional-tests",
+    "test:functional:dev:debug-server": "PORT=9090 babel-node scripts/functional-tests/debug-server",
     "test:argos": "argos upload functional-tests/screenshots --token $ARGOS_TOKEN || true",
-    "release": "./scripts/release.js",
-    "release:maintenance": "./scripts/release-maintenance.sh",
+    "release": "./scripts/release/publish.js",
+    "release:maintenance": "./scripts/release/publish-maintenance.sh",
     "release:view-changelog": "conventional-changelog -p angular"
   },
   "files": [
