+++ scripts/functional-tests/compile-watch.js
