+++ scripts/functional-tests/mocha.opts
@@ -1,0 +1,2 @@
+--compilers js:@babel/register
+@(src)/**/__tests__/**/*-test.js scripts/functional-tests/jsdom.js
