--- scripts/mocha.opts
@@ -1,2 +1,0 @@
---compilers js:@babel/register
-@(src)/**/__tests__/**/*-test.js scripts/jsdom.js
